fix: created a new file text_wrapper to have the functions to wrap the text withtin the terminal
